+++ src/entities/index.ts
@@ -1,1 +1,2 @@
-export * from './match';
+export { mockMatchData, mainBannerImg } from './match';
+export type { MatchData } from './match';
