+++ src/entities/match/model/index.ts
@@ -1,0 +1,1 @@
+export { mockMatchData } from './mock-data';
