+++ src/entities/match/ui/banner.ts
@@ -1,3 +1,3 @@
-export const MainBannerImg = {
+export const mainBannerImg = {
   topBanner: '/images/banners/main-top-banner.png',
 };
