+++ src/entities/match/ui/index.ts
@@ -1,0 +1,1 @@
+export { mainBannerImg } from './banner';
