+++ src/pages/home/index.ts
@@ -1,1 +1,1 @@
-export { HomePage } from './ui/HomePage';
+export { HomePage } from './ui';
