+++ src/widgets/date-picker/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export { DatePicker } from './ui';
