+++ src/widgets/filter-modal/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export { FilterModal } from './ui';
